REFACTOR: Plot func.
