+++ utils/plot_helper.py
@@ -44,7 +44,7 @@ def plot_feature_importance(feature_importance):
         feature_df = feature_df.iloc[:display_limit] # Show at most `display_limit` entries
         feature_df = feature_df.append(remainder)
 
-    feature_df["Feature_importance"] = feature_df["Feature_importance"].map('{:.3f}'.format)
+    feature_df["Feature_importance"] = feature_df["Feature_importance"].map('{:.3f}'.format).astype(np.float32)
     feature_df["Std"] = feature_df["Std"].map('{:.5f}'.format)
     feature_df_wo_links = feature_df.copy()
     feature_df["Name"] = feature_df["Name"].apply(lambda x: '<a href="https://www.ncbi.nlm.nih.gov/search/all/?term={}" title="Search on NCBI" target="_blank">{}</a>'.format(x, x)
@@ -285,34 +285,126 @@ def plot_pr_curve_cv(pr_curve_results, class_ratio_test, cohort_combos = None):
                     )
     return p
 
-# Perform EDA and Prepare Their Plots
+# Generate dendograms for hier. clustering
+def generate_dendrogram( matrix, labels, show_distances: bool = False, colorbar_title: str = "", ):
+    """Generate Dendrogram."""
+
+    # Initialize figure by creating upper dendrogram
+    fig = ff.create_dendrogram(
+        matrix,
+        orientation="bottom",
+        labels=labels,
+        colorscale=[gray_color] * 8,
+    )
+    for i in range(len(fig["data"])):
+        fig["data"][i]["yaxis"] = "y2"
+
+    # Create side dendrogram
+    dendro_side = ff.create_dendrogram(
+        matrix, orientation="right", colorscale=[gray_color] * 8
+    )
+    for i in range(len(dendro_side["data"])):
+        dendro_side["data"][i]["xaxis"] = "x2"
+
+    # Add Side Dendrogram Data to Figure
+    for data in dendro_side["data"]:
+        fig.add_trace(data)
+
+    # define dendro leaves
+    dendro_leaves = dendro_side["layout"]["yaxis"]["ticktext"]
+    dendro_leaves = list(map(int, dendro_leaves))
+
+    # get heatmap data (z)
+    if show_distances:
+        # calculate distances
+        heat_data = squareform(pdist(matrix))
+    else:
+        heat_data = matrix.values
+
+    # arrange the heatmap data according to the dendrogram clustering
+    heat_data = heat_data[dendro_leaves, :]
+    heat_data = heat_data[:, dendro_leaves]
+
+    heatmap = [
+        go.Heatmap(
+            x=dendro_leaves,
+            y=dendro_leaves,
+            z=heat_data,
+            colorscale=[
+                [0.0, blue_color],
+                [0.5, "#ffffff"],
+                [1.0, red_color],
+            ],
+            colorbar={"title": colorbar_title},
+            hovertemplate=(
+                "<b>Protein x:</b> %{x}<br><b>Protein y:</b> %{y}"
+                "<extra>r = %{z:.2f}</extra>"
+            ),
+        )
+    ]
+
+    heatmap[0]["x"] = fig["layout"]["xaxis"]["tickvals"]
+    heatmap[0]["y"] = dendro_side["layout"]["yaxis"]["tickvals"]
+    for data in heatmap:
+        fig.add_trace(data)
+
+    # modify layout
+    fig.update_layout(
+        {
+            "width": 800,
+            "height": 800,
+            "showlegend": False,
+            "hovermode": "closest",
+        }
+    )
+
+    # add labels to yaxis (needed for the hover)
+    fig["layout"]["yaxis"]["ticktext"] = fig["layout"]["xaxis"]["ticktext"]
+    fig["layout"]["yaxis"]["tickvals"] = fig["layout"]["xaxis"]["tickvals"]
+
+    # modify axes
+    params: dict = {
+        "mirror": False,
+        "showgrid": False,
+        "showline": False,
+        "zeroline": False,
+        "showticklabels": False,
+        "ticks": "",
+    }
+    fig.update_layout(
+        xaxis={"domain": [0.15, 1], **params},
+        xaxis2={"domain": [0, 0.15], **params},
+        yaxis={"domain": [0, 0.85], **params},
+        yaxis2={"domain": [0.825, 0.975], **params},
+    )
+
+    return fig
+
+# Perform EDA and Prepare their plots
 def perform_EDA(state):
     """
     Perform EDA on the dataset by given method and return the chart
     """
     
-    data = state.df_sub[state.proteins].astype('float').fillna(0)
+    data = state.df_sub[state.proteins].astype('float').fillna(0.0)
     if state.eda_method == "Hierarchical clustering":
-        columns = [x[:10] for x in data.columns] # to shorten the x-ticks
-        rows = list(data.index)
-        p = dash_bio.Clustergram(
-            data=data.loc[rows].values,
-            column_labels=columns,
-            row_labels=rows,
-            height=800,
-            width=800,
-            color_map=[
-                [0.0, blue_color],
-                [0.5, gray_color],
-                [1.0, red_color],
-            ],
-            cluster="all",
-            hidden_labels="row",
-            paper_bg_color='rgba(255,255,255,1)',
-            plot_bg_color='rgba(255,255,255,1)',
-            standardize="column"
+        data_to_be_correlated = data.iloc[:, state.data_range[0]:state.data_range[1]]
+        corr = data_to_be_correlated.corr(method="pearson")
+        labels = corr.columns
+        p = generate_dendrogram(
+            matrix=corr,
+            labels=labels,
+            colorbar_title="Pearson correlation coeff.",
         )
 
+        p.update_layout(autosize=True,
+                    width=800,
+                    height=800,
+                    xaxis_showgrid=False,
+                    yaxis_showgrid=False,
+                    plot_bgcolor= 'rgba(255, 255, 255, 0)',
+                    )
+        
     elif state.eda_method == "PCA":
         n_components = 2
         pca = PCA(n_components=n_components)
@@ -325,7 +417,7 @@ def perform_EDA(state):
             for i, var in enumerate(pca.explained_variance_ratio_ * 100)
         }
         labels["color"] = state.target_column
-        p = px.scatter(components, x=0, y=1, color=pca_color, labels=labels)
+        p = px.scatter(components, x=0, y=1, color=pca_color, labels=labels, hover_name=data.index)
         
         # Show feature lines
         if hasattr(state, "pca_show_features") and (state.pca_show_features==True):
